Allen relationships
This is an implementation of the Allen Relationships for Time Periods.
It uses literate programming using the lit tool. The generated output
is commited as well as the .lit source.

The documentaion is in Markdown and the code is Postgresql SQL. There
is only tests for the simple support functions.