schema qualify all calls to metadata functions.
The metadata functions should NOT require bitemporal_internal
on the search_path to work.

docs update to reflect renames & schema.

Tests updated and a check to make sure the schema is not on the
serach path before the other test run.

Renamed some functions to remove 'bitemporal' from the function name.
This should now be obvious given the need to schema qualify the calls.

Removed some excess whitespace.